Create BaseShape.
