+++ patterns/abstract_factory/tool_panel_factory/shapes/text_shape.dart
@@ -1,8 +1,8 @@
 import 'dart:ui';
 
-import '../pattern/shape.dart';
+import 'base_shape.dart';
 
-class TextShape extends Shape {
+class TextShape extends BaseShape {
   final String text;
   final double fontSize;
